fix(vector): count vectors against {prefixed}_meta; move IndexRow to service; sort indexes
